[TRL-461] chore: 일정 생성 관련 예외 - common 패키지로 이동
