+++ src/main/java/com/cosain/trilo/common/exception/place/InvalidCoordinateException.java
@@ -1,25 +1,25 @@
-package com.cosain.trilo.trip.domain.exception;
+package com.cosain.trilo.common.exception.place;
 
 import com.cosain.trilo.common.exception.CustomException;
 import org.springframework.http.HttpStatus;
 
-public class InvalidScheduleTitleException extends CustomException {
+public class InvalidCoordinateException extends CustomException {
 
-    private static final String ERROR_CODE = "schedule-0008";
+    private static final String ERROR_CODE = "place-0001";
     private static final HttpStatus HTTP_STATUS = HttpStatus.BAD_REQUEST;
 
-    public InvalidScheduleTitleException() {
+    public InvalidCoordinateException() {
     }
 
-    public InvalidScheduleTitleException(String debugMessage) {
+    public InvalidCoordinateException(String debugMessage) {
         super(debugMessage);
     }
 
-    public InvalidScheduleTitleException(Throwable cause) {
+    public InvalidCoordinateException(Throwable cause) {
         super(cause);
     }
 
-    public InvalidScheduleTitleException(String debugMessage, Throwable cause) {
+    public InvalidCoordinateException(String debugMessage, Throwable cause) {
         super(debugMessage, cause);
     }
