+++ src/main/java/com/cosain/trilo/common/exception/schedule/InvalidScheduleTitleException.java
@@ -1,25 +1,25 @@
-package com.cosain.trilo.trip.domain.exception;
+package com.cosain.trilo.common.exception.schedule;
 
 import com.cosain.trilo.common.exception.CustomException;
 import org.springframework.http.HttpStatus;
 
-public class InvalidCoordinateException extends CustomException {
+public class InvalidScheduleTitleException extends CustomException {
 
-    private static final String ERROR_CODE = "place-0001";
+    private static final String ERROR_CODE = "schedule-0008";
     private static final HttpStatus HTTP_STATUS = HttpStatus.BAD_REQUEST;
 
-    public InvalidCoordinateException() {
+    public InvalidScheduleTitleException() {
     }
 
-    public InvalidCoordinateException(String debugMessage) {
+    public InvalidScheduleTitleException(String debugMessage) {
         super(debugMessage);
     }
 
-    public InvalidCoordinateException(Throwable cause) {
+    public InvalidScheduleTitleException(Throwable cause) {
         super(cause);
     }
 
-    public InvalidCoordinateException(String debugMessage, Throwable cause) {
+    public InvalidScheduleTitleException(String debugMessage, Throwable cause) {
         super(debugMessage, cause);
     }
