+++ src/main/java/com/cosain/trilo/trip/domain/vo/Coordinate.java
@@ -1,6 +1,6 @@
 package com.cosain.trilo.trip.domain.vo;
 
-import com.cosain.trilo.trip.domain.exception.InvalidCoordinateException;
+import com.cosain.trilo.common.exception.place.InvalidCoordinateException;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
 import lombok.*;
