+++ src/main/java/com/cosain/trilo/trip/domain/vo/ScheduleTitle.java
@@ -1,6 +1,6 @@
 package com.cosain.trilo.trip.domain.vo;
 
-import com.cosain.trilo.trip.domain.exception.InvalidScheduleTitleException;
+import com.cosain.trilo.common.exception.schedule.InvalidScheduleTitleException;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
 import lombok.*;
