+++ src/test/java/com/cosain/trilo/unit/trip/application/schedule/service/schedule_create/ScheduleCreateServiceTest.java
@@ -1,10 +1,10 @@
 package com.cosain.trilo.unit.trip.application.schedule.service.schedule_create;
 
+import com.cosain.trilo.common.exception.schedule.NoScheduleCreateAuthorityException;
 import com.cosain.trilo.common.exception.schedule.TooManyDayScheduleException;
+import com.cosain.trilo.common.exception.schedule.TooManyTripScheduleException;
 import com.cosain.trilo.fixture.ScheduleFixture;
 import com.cosain.trilo.fixture.TripFixture;
-import com.cosain.trilo.trip.application.exception.NoScheduleCreateAuthorityException;
-import com.cosain.trilo.trip.application.exception.TooManyTripScheduleException;
 import com.cosain.trilo.trip.application.schedule.service.schedule_create.ScheduleCreateCommand;
 import com.cosain.trilo.trip.application.schedule.service.schedule_create.ScheduleCreateService;
 import com.cosain.trilo.trip.domain.entity.Day;
