+++ src/test/java/com/cosain/trilo/unit/trip/domain/vo/CoordinateTest.java
@@ -1,6 +1,6 @@
 package com.cosain.trilo.unit.trip.domain.vo;
 
-import com.cosain.trilo.trip.domain.exception.InvalidCoordinateException;
+import com.cosain.trilo.common.exception.place.InvalidCoordinateException;
 import com.cosain.trilo.trip.domain.vo.Coordinate;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.DisplayName;
@@ -9,7 +9,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 @Slf4j
 @DisplayName("[TripCommand] Coordinate 테스트")
